Created an addEvent method, and getEvent method
The add event method does not take all of the parameters that are required yet. Needs to be tested. addEvent will be called when the submit button is clicked on the add_event activity. The eventid will be the date and time of the event. For example, if the event is set to be 1/1/2020 at 5:00 PM, the eventID will be 1-1-2020-5-00PM, that way it is unique. Adding events should first check if an event already exists at that time, and not allow it if so.

The getEvent method will return a hashtable of every event for a particular date. This will need to be modified such that a list is built for each time and their respective title. Might need to use a hashtable as the value of the hashtable to store each time and respective event title.