Completed Basic version of app
Added a list view to the main activity that displays events, added a delete query, an edit query, as well as activities that show all of the information from each event rather than just their time and title. There is a slider of sorts that allows you to either edit or delete your event.